add author field to event, some other fixes

+++ event_planner/apps/users/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer
+from sqlalchemy import Column, String, BigInteger
 
 from ...db import Base
 
@@ -7,5 +7,5 @@ class User(Base):
     __tablename__ = 'users'
 
     username = Column(String, primary_key=True)
-    telegram_id = Column(Integer)
+    telegram_id = Column(BigInteger)
     hashed_password = Column(String)
